social_network/requirements.txt
antiorm==1.2.1
asgiref==3.4.1
db==0.1.1
dj-database-url==0.5.0
Django==3.2.8
django-jazzmin==2.4.8
django-on-heroku==1.1.2
djangorestframework==3.12.4
gunicorn==20.1.0
jsonfield==3.1.0
Markdown==3.2.1
Pillow==8.4.0
psycopg2-binary==2.9.2
pytz==2021.3
sqlparse==0.4.2
whitenoise==5.3.0
